Integrate CaptureSettings in UI and clean up UI
