+++ com.unity.render-pipelines.high-definition/Editor/Lighting/Reflection/HDProbeUI.Toolbar.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4a487934cbca92748ac7e4469d3cbebd
+guid: 1b036d2ab0d0e0842adcdbc254f3b942
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
