+++ com.unity.render-pipelines.high-definition/Editor/Lighting/Reflection/HDReflectionProbeEditor.cs
@@ -75,7 +75,6 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
         }
         
         SerializedObject m_AdditionalDataSerializedObject;
-        internal HDReflectionProbeUI m_UIState = new HDReflectionProbeUI();
 
         public bool sceneViewEditing
         {
@@ -95,16 +94,13 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             }
 
             base.OnEnable();
-
-            m_UIState.owner = this;
-            m_UIState.Reset(m_SerializedHDProbe, Repaint);
             
             InitializeTargetProbe();
 
             HDAdditionalReflectionData probe = (HDAdditionalReflectionData)m_AdditionalDataSerializedObject.targetObject;
             probe.influenceVolume.Init(probe);
 
-            //unhide previously hidden components
+            //unhide previously hidden components if any
             probe.hideFlags = HideFlags.None;
         }
     }
