--- com.unity.render-pipelines.high-definition/Editor/Lighting/Reflection/HDReflectionProbeEditorUtility.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c2498cd0a3473b54e8da8ce449481756
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
