+++ com.unity.render-pipelines.high-definition/Editor/Lighting/Reflection/HDReflectionProbeUI.cs
@@ -18,42 +18,37 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             toolBars = new[] { ToolBar.InfluenceShape | ToolBar.Blend | ToolBar.NormalBlend, ToolBar.CapturePosition };
         }
 
-        [Flags]
-        public enum Operation
-        {
-            None = 0,
-            FitVolumeToSurroundings = 1 << 0
-        }
-        Operation operations { get; set; }
+        //[Flags]
+        //public enum Operation
+        //{
+        //    None = 0,
+        //    FitVolumeToSurroundings = 1 << 0
+        //}
+        //Operation operations { get; set; }
 
-        public Editor owner { get; set; }
-        public bool HasOperation(Operation op) { return (operations & op) == op; }
-        public void ClearOperation(Operation op) { operations &= ~op; }
-        public void AddOperation(Operation op) { operations |= op; }
+        //public Editor owner { get; set; }
+        //public bool HasOperation(Operation op) { return (operations & op) == op; }
+        //public void ClearOperation(Operation op) { operations &= ~op; }
+        //public void AddOperation(Operation op) { operations |= op; }
 
-        public Matrix4x4 oldLocalSpace = Matrix4x4.identity;
+        //public Matrix4x4 oldLocalSpace = Matrix4x4.identity;
 
-        public bool HasAndClearOperation(Operation op)
-        {
-            var has = HasOperation(op);
-            ClearOperation(op);
-            return has;
-        }
+        //public bool HasAndClearOperation(Operation op)
+        //{
+        //    var has = HasOperation(op);
+        //    ClearOperation(op);
+        //    return has;
+        //}
 
-        public bool sceneViewEditing
-        {
-            get { return HDProbeUI.IsProbeEditMode(EditMode.editMode) && EditMode.IsOwner(owner); }
-        }
+        //public bool sceneViewEditing
+        //{
+        //    get { return HDProbeUI.IsProbeEditMode(EditMode.editMode) && EditMode.IsOwner(owner); }
+        //}
 
-        public override void Update()
-        {
-            operations = 0;
-            base.Update();
-        }
-
-        internal void UpdateOldLocalSpace(ReflectionProbe target)
-        {
-            oldLocalSpace = HDReflectionProbeEditorUtility.GetLocalSpace(target);
-        }
+        //public override void Update()
+        //{
+        //    operations = 0;
+        //    base.Update();
+        //}
     }
 }
