+++ com.unity.render-pipelines.high-definition/Editor/Lighting/Reflection/SerializedHDProbe.cs
@@ -6,11 +6,15 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
     {
         internal SerializedObject serializedObject;
         
+        internal SerializedProperty renderDynamicObjects;
+        internal SerializedProperty customBakedTexture;
+        //internal SerializedProperty timeSlicingMode;
+
         internal SerializedProperty proxyVolumeReference;
         internal SerializedProperty infiniteProjection;
 
         internal SerializedInfluenceVolume influenceVolume;
-
+        internal SerializedCaptureSettings captureSettings;
         internal SerializedFrameSettings frameSettings;
 
         internal SerializedProperty lightLayers;
@@ -32,9 +36,13 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
         internal SerializedHDProbe(SerializedObject serializedObject)
         {
             this.serializedObject = serializedObject;
+            
+            //customBakedTexture = serializedObject.Find((HDProbe p) => p.back);
+            //renderDynamicObjects = legacyProbe.FindProperty("m_RenderDynamicObjects");
 
             proxyVolumeReference = serializedObject.Find((HDProbe p) => p.proxyVolume);
             influenceVolume = new SerializedInfluenceVolume(serializedObject.Find((HDProbe p) => p.influenceVolume));
+            captureSettings = new SerializedCaptureSettings(serializedObject.Find((HDProbe p) => p.captureSettings));
             infiniteProjection = serializedObject.Find((HDProbe p) => p.infiniteProjection);
 
             frameSettings = new SerializedFrameSettings(serializedObject.Find((HDProbe p) => p.frameSettings));
@@ -51,6 +59,8 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
         {
             serializedObject.Update();
             //InfluenceVolume does not have Update. Add it here if it have in the future.
+            //CaptureSettings does not have Update. Add it here if it have in the future.
+            //FrameSettings does not have Update. Add it here if it have in the future.
         }
 
         internal virtual void Apply()
